Add venv to .gitignore; Add .python-version and requirements.txt; Refactor files structure; Rename files and move classes to their own files; Add username and password prompt;
