+++ README.md
@@ -10,9 +10,9 @@ We thank William Paivine for his initial work on this project.
 
 ## Installation
 
-To install the CLI, clone this repo. The CLI depends only on python's `requests` package, and requires python 3.6 or higher.
+To install the CLI, clone this repo. The CLI depends only on python's `requests` package, and requires python 3.7 or higher.
 
-To install the requests package, use `pip3 install requests`.
+To install the dependencies, use `pip install -r requirements.txt`.
 
 ## Basic Usage
 
@@ -22,7 +22,10 @@ For more details see [the guide](https://www.diderot.one/course/15/chapters/736/
 
 ### Credential Management
 
-Credentials are passed to the CLI in one of two ways. The first is to simply pass your credentials via the CLI using the `--username` and `--password` flags on the CLI. A more convenient way is to use a credentials file. The credential file format is simply a text file with your username on the first line and your password on the second. Use the --credentials argument to point the CLI towards a file containing your credentials. For easier usage, the CLI automatically looks at the file `~/private/.diderot/credentials` and then `~/.diderot/credentials` for a credentials file of this form. If this file exists, then the CLI will automatically log you in, and no credentials need to be explicitly provided to the CLI. An important note is that your credentials file must have only "owner can read and write" permissions. To do this, run `chmod 600 <credentials file>`.
+Credentials are passed to the CLI in one of three ways.
+The first is to simply pass your credentials via the CLI using the `--username` and `--password` flags on the CLI.
+A more convenient way is to use a credentials file. The credential file format is simply a text file with your username on the first line and your password on the second. Use the `--credentials` argument to point the CLI towards a file containing your credentials. For easier usage, the CLI automatically looks at the file `~/private/.diderot/credentials` and then `~/.diderot/credentials` for a credentials file of this form. If this file exists, then the CLI will automatically log you in, and no credentials need to be explicitly provided to the CLI. An important note is that your credentials file must have only "owner can read and write" permissions. To do this, run `chmod 600 <credentials file>`.
+If you omit the `--username`/`--password` pair **and** the `--credentials` flag, you will be prompted for them when running the command. The password is not echoed back to the terminal when typing.
 
 ## Student Version
