+++ diderot_admin
@@ -1,4 +1,8 @@
 #!/usr/bin/env python3
+import sys
+
 from standalone import DiderotAdmin
+
+assert sys.version_info >= (3, 7), "Python3.7 is required"
 
 DiderotAdmin().dispatch()
