--- standalone.py
@@ -1,576 +1,0 @@
-import argparse
-import datetime
-import json
-import os
-import shlex
-import sys
-from pathlib import Path
-
-from api_calls import DiderotAPIInterface
-from cli_utils import expand_file_path, print_list, APIError, BookNotFoundAPIError
-from constants import DEFAULT_CRED_LOCATIONS
-from models import Book, Chapter, Course, Lab, Part
-
-assert sys.version_info >= (3, 6), "Python3.6 is required"
-
-
-# Custom formatter for argparse which displays all positional arguments first.
-class Formatter(argparse.HelpFormatter):
-    # Use defined argument order to display usage.
-    def _format_usage(self, usage, actions, groups, prefix):
-        if prefix is None:
-            prefix = "usage: "
-
-        # If usage is specified, use that.
-        if usage is not None:
-            usage = usage % dict(prog=self._prog)
-
-        # If no optionals or positionals are available, usage is just prog.
-        elif usage is None and not actions:
-            usage = "%(prog)s" % dict(prog=self._prog)
-        elif usage is None:
-            prog = "%(prog)s" % dict(prog=self._prog)
-            # Build full usage string.
-            action_usage = self._format_actions_usage(actions, groups)  # NEW
-            usage = " ".join([s for s in [prog, action_usage] if s])
-            # Omit the long line wrapping code.
-        # Prefix with 'usage:'.
-        return "%s%s\n\n" % (prefix, usage)
-
-
-# exit_with_error prints an error message and exist with ret code 1.
-def exit_with_error(error_msg):
-    print(error_msg)
-    sys.exit(1)
-
-
-# Command line argument generator for the CLI.
-# Returns a parser that can be extended with additional commands.
-class DiderotCLIArgs(object):
-    # Generate a base parser object for use by both the user and admin CLI's.
-    @staticmethod
-    def generate_parser(prog, desc):
-        parser = argparse.ArgumentParser(prog=prog, description=desc, formatter_class=Formatter)
-        # Basic arguments not tied to a particular CLI version.
-        parser.add_argument("--url", default="https://api.diderot.one")
-        parser.add_argument("--username", default=None)
-        parser.add_argument("--password", default=None)
-        parser.add_argument(
-            "--credentials",
-            default=None,
-            help="Read credentials from a credentials file. \
-                                  The credential file format is a text file with the \
-                                  username on the first line and the password on the second.",
-        )
-        return parser
-
-    # Generate a parser for the user CLI.
-    @staticmethod
-    def generate_user_parser(prog, desc):
-        parser = DiderotCLIArgs.generate_parser(prog, desc)
-        subparsers = parser.add_subparsers(title="command", help="Different commands.", dest="command")
-        subparsers.required = True
-
-        # Subparser for list_assignments.
-        list_assignments = subparsers.add_parser(
-            "list_assignments", help="List all assignments in a course.", formatter_class=Formatter
-        )
-        list_assignments.add_argument("course", help="Course to list assignments of.")
-
-        # Subparser for list_courses.
-        subparsers.add_parser("list_courses", help="List all courses.", formatter_class=Formatter)
-
-        # Subparser for submit_assignment.
-        submit_assignment = subparsers.add_parser(
-            "submit_assignment", help="Submit handin to assignment.", formatter_class=Formatter
-        )
-        submit_assignment.add_argument("course", help="Course to submit assignment to.")
-        submit_assignment.add_argument("homework", help="Homework to submit to.")
-        submit_assignment.add_argument("handin_path", help="Path to handin file.")
-
-        # Subparser for download_assignment.
-        download_assignment = subparsers.add_parser(
-            "download_assignment", help="Download homework handout files.", formatter_class=Formatter
-        )
-        download_assignment.add_argument("course", help="Course to download assignment from.")
-        download_assignment.add_argument("homework", help="Homework assignment to download.")
-
-        return parser, subparsers
-
-    # Generate a parser for the admin CLI.
-    @staticmethod
-    def generate_admin_parser(prog, desc):
-        parser, subparsers = DiderotCLIArgs.generate_user_parser(prog, desc)
-
-        # Subparser for create_book.
-        # Example create_book <course label> --title <book title> --label <book label> 
-        create_book = subparsers.add_parser(
-            "create_book", help="Create a chapter in a book.", formatter_class=Formatter
-        )
-        create_book.add_argument("course", help="Course that the book belongs to.")
-
-        create_book.add_argument("--title", help="Optional title of new book (default = label)", default=None)
-        create_book.add_argument(
-            "--label", help="Label of new book"
-        )
-
-        # Subparser for create_chapter.
-        create_chapter = subparsers.add_parser(
-            "create_chapter", help="Create a chapter in a book.", formatter_class=Formatter
-        )
-        create_chapter.add_argument("course", help="Course that the book belongs to.")
-        create_chapter.add_argument("book", help="Book to create a chapter within.")
-        create_chapter.add_argument(
-            "--part",
-            help="Part number that the chapter belong within."
-        )
-        create_chapter.add_argument("--number", help="Number of the new chapter.", required=True)
-        create_chapter.add_argument("--title", help="Optional title of new chapter (default = Chapter)", default=None)
-        create_chapter.add_argument(
-            "--label", help="Optional label of new chapter (default= randomly generated)", default=None
-        )
-        create_chapter.add_argument(
-            "--publish_date",
-            help="Optional publish date for chapter in ISO format (yyyy-mm-dd hh:mm:ss+hh:mm)", default=None
-        )
-        create_chapter.add_argument(
-            "--due_date", help="Optional due date for chapter in ISO format (yyyy-mm-dd hh:mm:ss+hh:mm)", default=None
-        )
-
-        # Subparser for set_publish_date.
-        set_publish_date = subparsers.add_parser(
-            "set_publish_date", help="Set chapter publish date.", formatter_class=Formatter
-        )
-
-        set_publish_date.add_argument("course", help="Course to release chapter in.")
-        set_publish_date.add_argument("book", help="Book that the chapter belongs to.")
-        set_publish_date_chapter_id_group = set_publish_date.add_mutually_exclusive_group(required=True)
-        set_publish_date_chapter_id_group.add_argument(
-            "--chapter_number", default=None, help="Number of chapter to upload to.")
-        set_publish_date_chapter_id_group.add_argument(
-            "--chapter_label", default=None, help="Label of chapter to upload to.")
-        set_publish_date_group = set_publish_date.add_mutually_exclusive_group(required=True)
-        set_publish_date_group.add_argument(
-            "--publish_date", default=None, help="Publish date for chapter in ISO format (yyyy-mm-dd hh:mm:ss+hh:mm).")
-        set_publish_date_group.add_argument(
-            "--publish_on_week", default=None,
-            help="Publish on week from course start in format (week_num/week_day, hour:minute)")
-
-
-        # Subparser for create_part.
-        create_part = subparsers.add_parser("create_part", help="Create a part in a book", formatter_class=Formatter)
-        create_part.add_argument("course", help="Course that the book belongs to.")
-        create_part.add_argument("book", help="Book to create part in.")
-        create_part.add_argument("title", help="Title of the new part.")
-        create_part.add_argument("number", help="Number of the new part.")
-        create_part.add_argument("--label", help="Label for the new part (default = random).", default=None)
-
-        # Subparsers for release/unrelease chapter.
-        publish_chapter = subparsers.add_parser(
-            "publish_chapter", help="Release a book chapter", formatter_class=Formatter
-        )
-        publish_chapter.add_argument("course", help="Course to release chapter in.")
-        publish_chapter.add_argument("book", help="Book that the chapter belongs to.")
-        chapter_id_group = publish_chapter.add_mutually_exclusive_group(required=True)
-        chapter_id_group.add_argument("--chapter_number", default=None, help="Number of chapter to release.")
-        chapter_id_group.add_argument("--chapter_label", default=None, help="Label of chapter to release.")
-
-        retract_chapter = subparsers.add_parser(
-            "retract_chapter", help="Unrelease a book chapter", formatter_class=Formatter
-        )
-        retract_chapter.add_argument("course", help="Course to unrelease chapter in.")
-        retract_chapter.add_argument("book", help="Book that the chapter belongs to.")
-        chapter_id_group = retract_chapter.add_mutually_exclusive_group(required=True)
-        chapter_id_group.add_argument("--chapter_number", default=None, help="Number of chapter to unrelease.")
-        chapter_id_group.add_argument("--chapter_label", default=None, help="Label of chapter to unrelease.")
-
-        # Subparser for list_books.
-        list_books = subparsers.add_parser("list_books", help="List all books.", formatter_class=Formatter)
-        list_books.add_argument("course", help="Course label to filter book results by.", default=None, nargs="?")
-        list_books.add_argument("--all", action="store_true", help="Optional flag to show all books visible to user.")
-
-        # Subparser for list_chapters.
-        list_chapters = subparsers.add_parser(
-            "list_chapters", help="List all chapters in a book.", formatter_class=Formatter
-        )
-        list_chapters.add_argument("course", help="Course that the book belongs to.")
-        list_chapters.add_argument("book", help="Book to list chapters of.")
-        # TODO (rohany): add a filter by parts here.
-
-        # Subparser for list_parts.
-        list_parts = subparsers.add_parser("list_parts", help="List all parts of a book.", formatter_class=Formatter)
-        list_parts.add_argument("course", help="Course that the book belongs to.")
-        list_parts.add_argument("book", help="Book to list parts of.")
-
-        # Subparser for update_assignment.
-        update_assignment = subparsers.add_parser(
-            "update_assignment", help="Update the files linked to an assignment.", formatter_class=Formatter
-        )
-        update_assignment.add_argument("course", type=str, help="Course that assignment belongs to.")
-        update_assignment.add_argument("homework", type=str, help="Name of assignment to update.")
-        update_assignment.add_argument("--autograde-tar", type=str, help="Autograder tar for the assignment.")
-        update_assignment.add_argument("--autograde-makefile", type=str, help="Autograder Makefile for the assignment.")
-        update_assignment.add_argument("--writeup", type=str, help="PDF writeup for the assignment.")
-        update_assignment.add_argument("--handout", type=str, help="Student handout for the assignment.")
-
-        # Subparser for upload_chapter.
-        upload_chapter = subparsers.add_parser(
-            "upload_chapter", help="Upload content to a chapter.", formatter_class=Formatter
-        )
-        upload_chapter.add_argument("course", help="Course to upload chapter to.")
-        upload_chapter.add_argument("book", help="Book that the chapter belongs to.")
-        chapter_id_group = upload_chapter.add_mutually_exclusive_group(required=True)
-        chapter_id_group.add_argument("--chapter_number", default=None, help="Number of chapter to upload to.")
-        chapter_id_group.add_argument("--chapter_label", default=None, help="Label of chapter to upload to.")
-        file_type_group = upload_chapter.add_mutually_exclusive_group(required=True)
-        file_type_group.add_argument("--pdf", default=None, help="Upload PDF content.")
-        file_type_group.add_argument("--xml", default=None, help="Upload an XML/MLX document.")
-        upload_chapter.add_argument("--video_url", default=None, help="URL to a video to embed into the chapter.")
-        upload_chapter.add_argument("--xml_pdf", default=None, help="PDF version of the XML document for printing.")
-        upload_chapter.add_argument(
-            "--attach",
-            type=str,
-            nargs="+",
-            default=None,
-            help="A list of files, folders, or globs to upload as attachments for an XML/MLX document. \
-                                          Folders are recursively traversed.",
-        )
-
-        # Subparser for bulk book uploads.
-        upload_book = subparsers.add_parser(
-            "upload_book", help="Perform bulk upload of book content.", formatter_class=Formatter
-        )
-        upload_book.add_argument("course", help="Course that the book belongs to.")
-        upload_book.add_argument("upload_data", help="JSON upload details file.")
-
-        return parser, subparsers
-
-    # Verify argument information about both the admin and user CLI's.
-    @staticmethod
-    def validate_args(args):
-        password = args.password
-        if (password is None and args.username is not None) or (password is not None and args.username is None):
-            exit_with_error("Please supply both username and password")
-        if args.credentials is not None and (args.username is not None or password is not None):
-            exit_with_error("Cannot use a credentials file and input a username/password")
-
-
-# Class implementing the DiderotUser CLI.
-class DiderotUser(object):
-    def __init__(self, line=None):
-        parser, _ = DiderotCLIArgs.generate_user_parser("diderot", "User CLI interface for Diderot.")
-        if line is None:
-            self.args = parser.parse_args()
-        else:
-            self.args = parser.parse_args(shlex.split(line))
-        DiderotCLIArgs.validate_args(self.args)
-
-    def setup_client(self):
-        self.username = self.args.username
-        self.password = self.args.password
-        if self.username is None:
-            creds = DEFAULT_CRED_LOCATIONS
-            if self.args.credentials is not None:
-                creds = [self.args.credentials] + creds
-            for c in creds:
-                p = Path(c).expanduser()
-                if p.is_file():
-                    if (p.stat().st_mode & 0o177) != 0:
-                        exit_with_error(
-                            "Credentials file must have 0600 permissions. Run `chmod 600 <credentials>` first."
-                        )
-                    f = p.open("r")
-                    data = f.read().strip().split("\n")
-                    self.username = data[0]
-                    self.password = data[1]
-                    f.close()
-                    break
-                else:
-                    if c == self.args.credentials:
-                        exit_with_error("Input credentials path is invalid.")
-        if self.username is None:
-            exit_with_error("Supply your credentials via the CLI or a credentials file!")
-        self.api_client = DiderotAPIInterface(self.args.url)
-        self.api_client.login(self.username, self.password)
-
-    def dispatch(self):
-        commands = {
-            "download_assignment": self.download_assignment,
-            "list_assignments": self.list_assignments,
-            "list_courses": self.list_courses,
-            "submit_assignment": self.submit_assignment,
-        }
-        func = commands.get(self.args.command, None)
-        if func is None:
-            exit_with_error("Error parsing.")
-        try:
-            self.setup_client()
-            func()
-        except APIError as e:
-            exit_with_error(str(e))
-        self.api_client.client.close()
-
-    # For maintainability, keep the dispatch functions in alphabetical order.
-
-    def download_assignment(self):
-        self.api_client.download_assignment(self.args.course, self.args.homework)
-        print("Successfully downloaded assignment.")
-
-    def list_assignments(self):
-        course = Course(self.api_client.client, self.args.course)
-        labs = [hw["name"] for hw in Lab.list(course)]
-        if len(labs) == 0:
-            print("Course has no labs.")
-        else:
-            print_list(labs)
-
-    def list_courses(self):
-        print_list([c["label"] for c in Course.list(self.api_client.client)])
-
-    def submit_assignment(self):
-        try:
-            self.api_client.submit_assignment(self.args.course, self.args.homework, self.args.handin_path)
-            print("Assignment submitted successfully. Track your submission's status on Diderot.")
-        except APIError as e:
-            print(str(e))
-            exit_with_error("Something went wrong. Please try submitting on the Diderot website.")
-
-
-# Class implementing DiderotAdmin CLI.
-class DiderotAdmin(DiderotUser):
-    def __init__(self, line=None, sleep_time=5):
-        parser, _ = DiderotCLIArgs.generate_admin_parser("diderot_admin", "Admin CLI interface for Diderot.")
-        if line is None:
-            self.args = parser.parse_args()
-        else:
-            self.args = parser.parse_args(shlex.split(line))
-        self.sleep_time = sleep_time
-        DiderotCLIArgs.validate_args(self.args)
-
-    def dispatch(self):
-        commands = {
-            "create_book": self.create_book,
-            "create_chapter": self.create_chapter,
-            "create_part": self.create_part,
-            "list_books": self.list_books,
-            "list_chapters": self.list_chapters,
-            "list_parts": self.list_parts,
-            "publish_chapter": self.publish_chapter,
-            "set_publish_date": self.set_publish_date,
-            "retract_chapter": self.retract_chapter,
-            "update_assignment": self.update_assignment,
-            "upload_book": self.upload_book,
-            "upload_chapter": self.upload_chapter,
-        }
-        func = commands.get(self.args.command, None)
-        try:
-            if func is None:
-                super().dispatch()
-            else:
-                self.setup_client()
-                func()
-        except APIError as e:
-            exit_with_error(str(e))
-
-    # For maintainability, keep the dispatch functions in alphabetical order.
-
-    def create_book(self):
-        self.api_client.create_book(
-            self.args.course, self.args.title, self.args.label
-        )
-        print("Successfully created book.")
-
-    def create_chapter(self):
-        self.api_client.create_chapter(
-            self.args.course, self.args.book, self.args.part, self.args.number, self.args.title, self.args.label
-        )
-        print("Successfully created chapter.")
-
-    def create_part(self):
-        self.api_client.create_part(
-            self.args.course, self.args.book, self.args.number, self.args.title, self.args.label
-        )
-        print("Successfully created part.")
-
-    def list_books(self):
-        res = self.api_client.list_books(self.args.course, all=self.args.all)
-        print_list([c["label"] for c in res])
-
-    def list_chapters(self):
-        course = Course(self.api_client.client, self.args.course)
-        book = Book(course, self.args.book)
-        print_list(
-            [
-                "{}. {}".format(str(float(c["rank"])).rstrip("0").rstrip("."), c["title"])
-                for c in Chapter.list(course, book)
-            ]
-        )
-
-    def list_parts(self):
-        course = Course(self.api_client.client, self.args.course)
-        book = Book(course, self.args.book)
-        print_list(["{}. {}".format(c["rank"], c["title"]) for c in Part.list(course, book)])
-
-    def publish_chapter(self):
-        self.api_client.release_unrelease_chapter(self.args.course, self.args.book, self.args, release=True)
-        print("Success publishing chapter.")
-
-    def set_publish_date(self):
-        self.api_client.set_publish_date(self.args.course, self.args.book, self.args)
-        print("Successfully set publish date for the chapter.")
-
-    def retract_chapter(self):
-        self.api_client.release_unrelease_chapter(self.args.course, self.args.book, self.args, release=False)
-        print("Success retracting chapter.")
-
-    def update_assignment(self):
-        try:
-            self.api_client.update_assignment(self.args.course, self.args.homework, self.args)
-            print("Success uploading files.")
-        except APIError as e:
-            print(str(e))
-            exit_with_error("Uploading files failed. Try using the Web UI.")
-
-    def upload_book(self):
-        file_path = expand_file_path(self.args.upload_data)
-        if not os.path.exists(file_path):
-            exit_with_error("input file path {} does not exist".format(file_path))
-        with open(file_path, "rb") as schema:
-            try:
-                book_data = json.load(schema)
-            except Exception as e:
-                exit_with_error("Failed loading json schema with error: {}".format(e))
-
-        def get_or_none(obj, field):
-            if field in obj:
-                return obj[field]
-            return None
-
-        file_prefix = os.path.dirname(file_path)
-
-        def adjust_search_path(path):
-            if path is None:
-                return None
-            return os.path.join(file_prefix, path)
-
-        # Collect the necessary Diderot objects.
-        course = Course(self.api_client.client, self.args.course)
-        book_label = get_or_none(book_data, "book")
-        # Try out "label", more consistent with Diderot terminology
-        if book_label is None:
-            book_label = get_or_none(book_data, "label")
-        # If book label is still None, then error out.
-        if book_label is None:
-            exit_with_error("Please specify a valid book to upload into")
-
-        book_title = book_data.get("title", book_label)
-
-        try:
-            book = Book(course, book_label)
-        except BookNotFoundAPIError:
-            Book.create(course, book_title, book_label)
-            book = Book(course, book_label)
-
-        book_data_chapters = book_data.get("chapters", [])
-
-        book_data_parts = book_data.get("parts", [])
-        book_data_part_numbers = set([c.get("number") for c in book_data_parts])
-        chapters_data_part_numbers = set([c.get("part") for c in book_data_chapters])
-        actual_part_numbers = set([int(float(c["rank"])) for c in Part.list(course, book)])
-        union_part_numbers = actual_part_numbers.union(book_data_part_numbers)
-        if union_part_numbers != set(range(1, len(union_part_numbers) + 1)):
-            exit_with_error(f"invalid JSON: resulting parts numbers are inconsistent, "
-                            f"should be a sequence of integers starting with 1 including existing parts. "
-                            f"Current numbers set is: {actual_part_numbers} and resulting using json "
-                            f"is {union_part_numbers}")
-        elif not chapters_data_part_numbers.issubset(union_part_numbers):
-            exit_with_error(f"invalid JSON: some parts numbers for chapters are invalid. "
-                            f"Resulting part number set (existing and new) is {union_part_numbers} and specified in"
-                            f" chapter number set is {chapters_data_part_numbers}")
-
-        book_data_chapter_numbers = set([c.get("number") for c in book_data_chapters])
-        actual_chapter_numbers = set([int(float(c["rank"])) for c in Chapter.list(course, book)])
-        union_chapter_numbers = actual_chapter_numbers.union(book_data_chapter_numbers)
-        if union_chapter_numbers != set(range(1, len(union_chapter_numbers)+1)):
-            exit_with_error(f"invalid JSON: resulting chapters numbers are inconsistent, "
-                            f"should be a sequence of integers starting with 1 including existing chapters. "
-                            f"Current numbers set is: {actual_chapter_numbers} and resulting using json "
-                            f"is {union_chapter_numbers}")
-
-        # If the upload contains parts, create them.
-        parts = get_or_none(book_data, "parts")
-        if parts is not None:
-            for part in parts:
-                if not Part.exists(course, book, get_or_none(part, "number")):
-                    self.api_client.create_part(
-                        course.label,
-                        book.label,
-                        get_or_none(part, "number"),
-                        get_or_none(part, "title"),
-                        get_or_none(part, "label"),
-                    )
-
-        # Upload and maybe create the chapters in the input.
-        chapters = get_or_none(book_data, "chapters")
-        if chapters is None:
-            exit_with_error("invalid JSON: could not find field 'chapters'")
-
-        for chapter in chapters:
-            # Extract data from chapter json
-            number = get_or_none(chapter, "number")
-            label = get_or_none(chapter, "label")
-            title = get_or_none(chapter, "title")
-            attachments = get_or_none(chapter, "attachments")
-
-            self.args.book = book_label
-            self.args.part_num = get_or_none(chapter, "part")
-            self.args.chapter_number = number
-            self.args.chapter_label = label
-            self.args.pdf = adjust_search_path(get_or_none(chapter, "pdf"))
-            self.args.video_url = adjust_search_path(get_or_none(chapter, "video"))
-            self.args.xml = adjust_search_path(get_or_none(chapter, "xml"))
-            self.args.xml_pdf = adjust_search_path(get_or_none(chapter, "xml_pdf"))
-            self.args.publish_date = get_or_none(chapter, "publish_on_date")
-            self.args.publish_on_week = get_or_none(chapter, "publish_on_week")
-
-
-            if number is None:
-                exit_with_error(f"invalid JSON: must provide field 'number' for chapter {chapter}")
-
-            if Chapter.exists(course, book, number):
-                self.set_publish_date()
-            else:
-                # create chapter
-                if self.args.part_num is None:
-                    exit_with_error("Chapter creation in a book requires 'part' field for chapters")
-                self.api_client.create_chapter(
-                    course.label, book.label, self.args.part_num, number, title, label, self.args.publish_date, self.args.publish_on_week)
-                print(f"Successfully created chapter number ({number}), label ({label}, title ({title}).")
-
-            # Upload the target files to the chapter now.
-            self.args.attach = None
-            if attachments is not None:
-                self.args.attach = [adjust_search_path(path) for path in attachments]
-            # Set default arguments that we wont use, but upload_chapter expects.
-            print(f"Uploading chapter number: {number}...")
-            try:
-                self.api_client.upload_chapter(
-                    course.label, book.label, number, None, self.args, sleep_time=self.sleep_time,
-                )
-            except APIError as e:
-                print(e)
-                exit_with_error("Failure uploading chapter. Aborting")
-            print("Successfully uploaded chapter.")
-
-    def upload_chapter(self):
-        if self.args.attach is not None and self.args.xml is None:
-            exit_with_error("Cannot use --attach if not uploading xml/mlx.\nFailure uploading chapter.")
-        self.api_client.upload_chapter(
-            self.args.course,
-            self.args.book,
-            self.args.chapter_number,
-            self.args.chapter_label,
-            self.args,
-            sleep_time=self.sleep_time,
-        )
-        print("Chapter uploaded successfully.")
