+++ standalone/__init__.py
@@ -1,0 +1,4 @@
+from .diderot_admin import *
+from .diderot_api import *
+from .diderot_cli_args import *
+from .diderot_user import *
